requirements.txt
# Manipulação de dados
pandas>=2.0.0
numpy>=1.24.0

# Requisições HTTP
requests>=2.31.0

# Aprendizado de máquina
scikit-learn>=1.4.0
joblib>=1.3.0

# Visualização (opcional, para gráficos e análises)
matplotlib>=3.8.0
seaborn>=0.13.0

# Logs e utilitários
python-dotenv>=1.0.0
